+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/messenger/compose/view/room/settings/TimelineElementMetadata.kt
@@ -8,8 +8,10 @@ import androidx.compose.foundation.layout.ColumnScope
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.paddingFromBaseline
 import androidx.compose.foundation.layout.size
@@ -20,11 +22,14 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Switch
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.derivedStateOf
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Alignment.Companion.CenterVertically
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.style.TextOverflow
@@ -42,15 +47,21 @@ import de.connect2x.messenger.compose.view.common.TooltipText
 import de.connect2x.messenger.compose.view.get
 import de.connect2x.messenger.compose.view.i18n.I18nView
 import de.connect2x.messenger.compose.view.room.timeline.DateStickyHeader
+import de.connect2x.messenger.compose.view.room.timeline.element.TimelineElementHolder
 import de.connect2x.messenger.compose.view.room.timeline.element.TimelineElementViewSelector
 import de.connect2x.messenger.compose.view.room.timeline.element.util.Tooltip
+import de.connect2x.messenger.compose.view.util.waitForElementWithTimeout
 import de.connect2x.trixnity.messenger.viewmodel.UserInfoElement
 import de.connect2x.trixnity.messenger.viewmodel.room.settings.TimelineElementMetadataViewModel
 import de.connect2x.trixnity.messenger.viewmodel.room.timeline.elements.TimelineElementHolderViewModel
+import de.connect2x.trixnity.messenger.viewmodel.room.timeline.elements.TimelineElementViewModel
 import de.connect2x.trixnity.messenger.viewmodel.util.EventReactions
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.flow.filterNotNull
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
 import net.folivo.trixnity.core.model.UserId
 import kotlin.time.Duration.Companion.milliseconds
-
 
 interface TimelineElementMetadataView {
     @Composable
@@ -76,26 +87,49 @@ class TimelineElementMetadataViewImpl : TimelineElementMetadataView {
     ) {
         val i18n = DI.get<I18nView>()
 
-        val elementHistory = viewModel.elementHistory.collectAsState().value
-        val element = viewModel.element.collectAsState().value
+        val timelineElementViewSelector = DI.get<TimelineElementViewSelector>()
+        var elementHistory by remember { mutableStateOf(listOf<TimelineElementHolderViewModel>()) }
+        var element by remember { mutableStateOf<TimelineElementHolderViewModel?>(null) }
         val sender = element?.sender?.collectAsState()?.value
         val reactions = element?.reactions?.collectAsState()?.value
         val readers = element?.readers?.collectAsState()?.value
-        if (element == null || reactions == null || readers == null || sender == null) {
-            LoadingSpinner(Modifier.fillMaxSize())
-        } else {
-            ExtrasPaneHeader(
-                title = i18n.timelineElementMetadataTitle(),
-                error = null,
-                onBack = { viewModel.back() },
-                backButtonType = if (isSinglePane || isBottomOfStack.not()) BACK else CLOSE,
-            ) {
+
+        LaunchedEffect(Unit) {
+            launch {
+                viewModel.elementHistory.filterNotNull().collect { history ->
+                    withContext(Dispatchers.Default) {
+                        history.forEach { element ->
+                            launch {
+                                waitForElementWithTimeout(timelineElementViewSelector, element)
+                            }
+                        }
+                    }
+                    elementHistory = history
+                }
+            }
+            viewModel.element.filterNotNull().collect { newElement ->
+                waitForElementWithTimeout(timelineElementViewSelector, newElement)
+                element = newElement
+            }
+        }
+
+        ExtrasPaneHeader(
+            title = i18n.timelineElementMetadataTitle(),
+            error = null,
+            onBack = { viewModel.back() },
+            backButtonType = if (isSinglePane || isBottomOfStack.not()) BACK else CLOSE,
+        ) {
+            if (reactions == null || readers == null || sender == null || element == null || elementHistory.isEmpty()) {
+                LoadingSpinner(Modifier.fillMaxSize())
+            } else {
                 Box(
                     Modifier.fillMaxSize()
                 ) {
                     Column(
-                        Modifier
+                        verticalArrangement = Arrangement.Top,
+                        modifier = Modifier
                             .padding(PaddingValues(vertical = 0.dp, horizontal = 20.dp))
+                            .fillMaxSize()
                     ) {
                         SubHeading(i18n.timelineElementMetadataSender())
                         UserInfo(
@@ -103,11 +137,17 @@ class TimelineElementMetadataViewImpl : TimelineElementMetadataView {
                             onOpenUserProfile = viewModel::openUserProfile,
                         )
                         SubHeading(i18n.timelineElementMetadataMessage())
-                        MessageContentHistorySwitch(element, elementHistory)
+                        Column(modifier = Modifier.fillMaxWidth().fillMaxHeight(0.6f)) {
+                            element?.let {
+                                MessageContentHistorySwitch(it, elementHistory)
+                            }
+                        }
                         SmallSpacer()
                         HorizontalDivider()
                         MiddleSpacer()
-                        ReadersAndReactions(element, viewModel)
+                        element?.let {
+                            ReadersAndReactions(it, viewModel)
+                        }
                         SmallSpacer()
                     }
                 }
@@ -138,7 +178,7 @@ fun ColumnScope.ReadersAndReactions(
         val allReadersAndReactions = remember(readers, reactions) {
             (readers.associate { it.userId to EventReactions.ByUserInfo(mapOf(), it, false) } +
                     reactions.byUser).values.toList()
-        }
+        }.sortedByDescending { it.reactions.size }
         val hasReadersOrReactions = allReadersAndReactions.isNotEmpty()
 
         if (hasReadersOrReactions) {
@@ -147,13 +187,14 @@ fun ColumnScope.ReadersAndReactions(
                 style = MaterialTheme.typography.titleMedium,
             )
             SmallSpacer()
-            LazyColumn(verticalArrangement = Arrangement.spacedBy(10.dp)) {
-                items(allReadersAndReactions, { it.sender.userId.full }) { eventReaction ->
+            LazyColumn {
+                items(allReadersAndReactions) { eventReaction ->
                     UserInfo(
                         eventReaction.sender,
                         eventReaction.reactions.keys,
                         onOpenUserProfile = viewModel::openUserProfile,
                     )
+                    Spacer(Modifier.height(5.dp))
                 }
             }
         } else {
@@ -230,12 +271,12 @@ private fun UserInfo(
 @Composable
 private fun ColumnScope.MessageContentHistorySwitch(
     element: TimelineElementHolderViewModel,
-    elementHistory: List<TimelineElementHolderViewModel>?,
+    elementHistory: List<TimelineElementHolderViewModel>,
 ) {
     val i18n = DI.get<I18nView>()
     var showHistory by remember { mutableStateOf(false) }
 
-    if (elementHistory.isNullOrEmpty().not() && elementHistory.size > 1) {
+    if (elementHistory.isNotEmpty() && elementHistory.size > 1) {
         Row(
             verticalAlignment = CenterVertically,
             modifier = Modifier.clickable { showHistory = showHistory.not() }.buttonPointerModifier(),
@@ -253,32 +294,55 @@ private fun ColumnScope.MessageContentHistorySwitch(
     if (showHistory) {
         MessageHistory(elementHistory)
     } else {
+        DateStickyHeader(element.formattedDate)
+        Spacer(Modifier.height(8.dp))
         MessageContent(element)
     }
 }
 
 @Composable
-private fun ColumnScope.MessageContent(
-    messageHolder: TimelineElementHolderViewModel?,
-) {
-    messageHolder?.let { holder ->
-        DateStickyHeader(messageHolder.formattedDate)
-        holder.element.collectAsState().value?.let { element ->
-            Column(
-                Modifier.padding(end = 8.dp),
-            ) {
-                DI.get<TimelineElementViewSelector>().createAsPreview(holder, element)
-            }
-            SmallSpacer()
+private fun ColumnScope.MessageContent(messageHolder: TimelineElementHolderViewModel) {
+    val element = messageHolder.element.collectAsState().value
+    val timelineElementViewSelector = DI.get<TimelineElementViewSelector>()
+    Column {
+        element?.let { element ->
+            timelineElementViewSelector.createAsPreview(messageHolder, element)
         }
     }
 }
 
 @Composable
-private fun ColumnScope.MessageHistory(elementHistory: List<TimelineElementHolderViewModel>?) {
-    if (elementHistory?.isNotEmpty() == true) {
-        elementHistory.forEach { elementHolder ->
-            MessageContent(elementHolder)
+private fun ColumnScope.MessageHistory(elementHistory: List<TimelineElementHolderViewModel>) {
+    if (elementHistory.isNotEmpty()) {
+        val elementHistoryGrouped by derivedStateOf {
+            buildList(elementHistory.size) {
+                var lastDate: String? = null
+                for (index in elementHistory.indices) {
+                    val vm = elementHistory[index]
+                    when {
+                        lastDate == vm.formattedDate -> add(null to vm)
+                        vm.element.value is TimelineElementViewModel.Empty -> add(null to vm)
+                        else -> {
+                            add(vm.formattedDate to vm)
+                            lastDate = vm.formattedDate
+                        }
+                    }
+                }
+            }
+        }
+
+        LazyColumn {
+            elementHistoryGrouped.forEach { (date, viewModel) ->
+                if (date != null) {
+                    item("date-$date-${viewModel.key}") {
+                        DateStickyHeader(date)
+                        Spacer(Modifier.height(8.dp))
+                    }
+                }
+                item(viewModel.key) {
+                    MessageContent(viewModel)
+                }
+            }
         }
     }
 }