+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Check for server reachability when logout fails
 - Make room notification settings collapsable and change setting design to the one used by other room settings
 - Disable notification setting sub-options when notifications are disabled
+- Sort users with reactions on the top of the user list on the metadata view
+- Save timeline scroll state when opening sub-menu of room
 - Use fastzip in CI
 - Upgrade Trixnity to 4.13.2
 
@@ -28,6 +30,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Align padding for sticky date header with dates in the timeline on mobile
+- Date bubble is only shown when the date is changed in message metadata view
+- Fix content of message replacements in history of message metadata view
+- Allow to show audio messages as preview
+- Scrollbar to message metadata view for long messages/long message history
 - Inserting Mentions fails for multiple ones
 
 ### Security